Fix resequence ID feature so that it works with unified networks
The resequence ID feature gives a human readable ID to every node/edge
that is scoped to the exported data. Some users find this useful.

Previously the IDs were being generated prior to the processing of the
"merge networks" option, meaning that the IDs were not unique across
the entire exported data set. This commit fixes that.